instaintel-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
pandas
numpy
scikit-learn
prophet==1.1.5
celery==5.3.6
pytest
httpx